backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
openai
python-dotenv
pytest
